requirements.txt
flask
spotipy
onnxruntime
opencv-python
numpy
python-dotenv